chore: add missing licenses and repos for 21 grammars
Fill in license and repo fields for grammars that had empty values:
- d, agda, commonlisp, erlang, fish, ruby (MIT, Apache-2.0, Unlicense)
- cmake, dot, graphql, ninja, ssh-config (MIT)
- kdl (original, MIT OR Apache-2.0)
- query (Apache-2.0)
- ada, julia, matlab, sparql, elisp, typst (MIT)
- textproto (ISC)
- uiua (MPL-2.0)

This brings the umbrella crate from 77 to 98 grammars.
