+++ langs/group-fern/agda/def/arborium.kdl
@@ -1,6 +1,6 @@
-repo ""
+repo "https://github.com/tree-sitter/tree-sitter-agda"
 commit ""
-license ""
+license "MIT"
 
 grammar {
     id "agda"
