+++ langs/group-maple/kdl/def/arborium.kdl
@@ -1,6 +1,6 @@
-repo ""
+repo "https://github.com/bearcove/arborium"
 commit ""
-license ""
+license "MIT OR Apache-2.0"
 
 grammar {
     id "kdl"
