+++ langs/group-moss/matlab/def/arborium.kdl
@@ -1,6 +1,6 @@
-repo ""
+repo "https://github.com/acristoffers/tree-sitter-matlab"
 commit ""
-license ""
+license "MIT"
 
 grammar {
     id "matlab"
