+++ langs/group-sage/elisp/def/arborium.kdl
@@ -1,6 +1,6 @@
-repo ""
+repo "https://github.com/Wilfred/tree-sitter-elisp"
 commit ""
-license ""
+license "MIT"
 
 grammar {
     id "elisp"
