+++ packages/arborium/src/loader.ts
@@ -8,8 +8,12 @@
  * 3. Parse and highlight using the grammar's tree-sitter parser
  */
 
-import type { ParseResult, ArboriumConfig, Grammar, Session, Span } from "./types.js";
+import type { ParseResult, ArboriumConfig, Grammar, Session } from "./types.js";
 import { availableLanguages, pluginVersion } from "./plugins-manifest.js";
+import { spansToHtml, escapeHtml } from "./utils.js";
+
+// Re-export utilities
+export { spansToHtml } from "./utils.js";
 
 // Default config
 export const defaultConfig: Required<ArboriumConfig> = {
@@ -361,91 +365,6 @@ export async function loadGrammar(
       // No-op for now, plugins are cached
     },
   };
-}
-
-/** Convert spans to HTML */
-export function spansToHtml(source: string, spans: Span[]): string {
-  // Sort spans by start position
-  const sorted = [...spans].sort((a, b) => a.start - b.start);
-
-  let html = "";
-  let pos = 0;
-
-  for (const span of sorted) {
-    // Skip overlapping spans
-    if (span.start < pos) continue;
-
-    // Add text before span
-    if (span.start > pos) {
-      html += escapeHtml(source.slice(pos, span.start));
-    }
-
-    // Get tag for capture
-    const tag = getTagForCapture(span.capture);
-    const text = escapeHtml(source.slice(span.start, span.end));
-
-    if (tag) {
-      html += `<a-${tag}>${text}</a-${tag}>`;
-    } else {
-      html += text;
-    }
-
-    pos = span.end;
-  }
-
-  // Add remaining text
-  if (pos < source.length) {
-    html += escapeHtml(source.slice(pos));
-  }
-
-  return html;
-}
-
-/** Get the short tag for a capture name */
-function getTagForCapture(capture: string): string | null {
-  if (capture.startsWith("keyword") || capture === "include" || capture === "conditional") {
-    return "k";
-  }
-  if (capture.startsWith("function") || capture.startsWith("method")) {
-    return "f";
-  }
-  if (capture.startsWith("string") || capture === "character") {
-    return "s";
-  }
-  if (capture.startsWith("comment")) {
-    return "c";
-  }
-  if (capture.startsWith("type")) {
-    return "t";
-  }
-  if (capture.startsWith("variable")) {
-    return "v";
-  }
-  if (capture.startsWith("number") || capture === "float") {
-    return "n";
-  }
-  if (capture.startsWith("operator")) {
-    return "o";
-  }
-  if (capture.startsWith("punctuation")) {
-    return "p";
-  }
-  if (capture.startsWith("tag")) {
-    return "tg";
-  }
-  if (capture.startsWith("attribute")) {
-    return "at";
-  }
-  return null;
-}
-
-/** Escape HTML special characters */
-function escapeHtml(text: string): string {
-  return text
-    .replace(/&/g, "&amp;")
-    .replace(/</g, "&lt;")
-    .replace(/>/g, "&gt;")
-    .replace(/"/g, "&quot;");
 }
 
 /** Get current config, optionally merging with overrides */
