+++ xtask/templates/packages_arborium_package.stpl.json
@@ -7,7 +7,8 @@
   "devDependencies": {
     "typescript": "^5.7.2",
     "vite": "^7.2.6",
-    "vite-plugin-dts": "^4.5.0"
+    "vite-plugin-dts": "^4.5.0",
+    "vitest": "^4.0.16"
   },
   "exports": {
     ".": {
@@ -82,6 +83,8 @@
     "gen:manifest": "cd ../.. && cargo xtask gen-manifest",
     "prepublishOnly": "npm run gen:manifest && npm run build && test -f dist/arborium.iife.js || (echo 'ERROR: dist/arborium.iife.js missing!' && exit 1)",
     "preview": "vite preview",
+    "test": "vitest run",
+    "test:watch": "vitest",
     "typecheck": "tsc --noEmit"
   },
   "type": "module",
